tests

+++ src/app/db/daos/playlist_entry.hpp
@@ -1,11 +1,9 @@
 #pragma once
 
-#include <string>
-
 namespace firelight::db {
 struct PlaylistEntry {
-  long playlistId;
-  long libraryEntryId;
+  int playlistId;
+  int libraryEntryId;
 };
 
 } // namespace firelight::db
